+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Version 0.0.6 *(WIP)*
 
 * Create dedicated object for each operation in a document([#3](https://github.com/yshrsmz/kgql/issues/3))
 * Change KgqlRequestBody & KgqlResponse to interface([#8](https://github.com/yshrsmz/kgql/issues/8))
+* Change suffix of generate classes to Document, from DocumentWrapper
 
 
 Version 0.0.5 *(2019-01-15)*